Rename functions in the compiled module (#14)
Functions are renamed to include argument types. Cassette overdubs are handled specially to include the base function called. That helps with debugging, so not every function is overdub_xxxxxxx.

This pulls in more code from CUDAnative to support renaming. It also removes some duplicated methods in the compiled LLVM module.